policy: inline parser feature profile owner module
