+++ Cargo.toml
@@ -43,7 +43,6 @@ members = [
   "crates/bdd-scenario-core",
   "crates/bdd-grid-core",
   "crates/feature-policy-core",
-  "crates/parser-feature-profile-core",
   "crates/governance-status-core",
   "crates/concurrency-bounded-map-core",
   "crates/concurrency-caps-contract-core",
