--- crates/parser-feature-profile-core/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "adze-parser-feature-profile-core"
-version = "0.1.0"
-edition.workspace = true
-rust-version.workspace = true
-license = "MIT OR Apache-2.0"
-description = "Single-responsibility parser feature profile and backend resolution helpers."
-repository = "https://github.com/EffortlessMetrics/adze"
-publish = ["crates.io"]
-keywords = ["parser", "feature-flags", "tree-sitter", "glr"]
-categories = ["parsing", "development-tools"]
-homepage = "https://github.com/EffortlessMetrics/adze"
-documentation = "https://docs.rs/adze-parser-feature-profile-core"
-
-[features]
-default = []
-pure-rust = ["adze-parser-backend-core/pure-rust"]
-tree-sitter-c2rust = ["adze-parser-backend-core/tree-sitter-c2rust"]
-tree-sitter-standard = ["adze-parser-backend-core/tree-sitter-standard"]
-glr = ["pure-rust", "adze-parser-backend-core/glr"]
-strict_docs = []
-strict_api = []
-
-[dependencies]
-adze-parser-backend-core = { version = "0.1.0", path = "../parser-backend-core" }
